Se implementa CRUD de Planes de Mantenimiento.

+++ app/Http/Controllers/PlanMantoController.php
@@ -4,15 +4,37 @@ namespace App\Http\Controllers;
 
 use App\Models\PlanManto;
 use Illuminate\Http\Request;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\DB;
+use Hashids\Hashids;
 
 class PlanMantoController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    protected $hashids;
+    public function __construct()
     {
-        $planes = PlanManto::all();
+        $this->hashids = new Hashids(env('APP_KEY'), 10);
+    }
+    public function index(Request $request)
+    {
+
+        $query = PlanManto::query();
+
+        if ($request->has('search')) {
+            $searchTerm = $request->search;
+            $query->where('nombre', 'LIKE', "%{$searchTerm}%")
+                ->orWhere('id_plan_manto', 'LIKE', "%{$searchTerm}%");
+        }
+        $planes = $query->paginate(10);
+
+        $planes->getCollection()->transform(function ($plan) {
+            $plan->hashed_id = $this->hashids->encode($plan->id_plan_manto);
+            return $plan;
+        });
+        // Retorna la vista 'producto.index' con los productos codificados
         return view('plan_manto.index', compact('planes'));
     }
 
@@ -29,15 +51,23 @@ class PlanMantoController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $request->validate([
             'nombre' => 'required|string|max:25|unique:plan_manto',
             'descripcion' => 'required|string|max:45|unique:plan_manto',
             'frecuencia_mes' => 'required|integer|min:1',
         ]);
 
-        PlanManto::create($validatedData);
+        $currentDateTime = Carbon::now('America/Guatemala');
 
-        return redirect()->route('plan_manto.index')->with('success', 'Plan de mantenimiento creado exitosamente.');
+        $planes = DB::table('plan_manto')->insertGetId([
+            'nombre' => $request->nombre,
+            'descripcion' => $request->descripcion,
+            'frecuencia_mes' => $request->frecuencia_mes,
+            'created_at' => $currentDateTime,
+            'updated_at' => $currentDateTime,
+        ]);
+
+        return redirect()->route('planes.index')->with('success', 'Plan de mantenimiento creado exitosamente.');
     }
 
     /**
@@ -51,34 +81,56 @@ class PlanMantoController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(PlanManto $planManto)
+    public function edit($hashedId)
     {
-        return view('plan_manto.edit', compact('planManto'));
+        //Se decodifica el ID encriptado
+        $id_plan_manto = $this->hashids->decode($hashedId)[0] ?? null;
+        if (!$id_plan_manto) {
+            abort(404);
+        }
+        //Se busca el ID en la tabla para validar
+        $planes = PlanManto::findOrFail($id_plan_manto);
+        $planes->hashed_id = $hashedId; // Usamos el hashed_id directamente nuevamente
+        return view('plan_manto.edit', compact('planes'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, PlanManto $planManto)
+    public function update(Request $request, $hashedId)
     {
-        $validatedData = $request->validate([
+        $id_plan_manto = $this->hashids->decode($hashedId)[0] ?? null;
+        if (!$id_plan_manto) {
+            abort(404);
+        }
+        $planManto = PlanManto::findOrFail($id_plan_manto);
+        $request->validate([
             'nombre' => 'required|string|max:25|unique:plan_manto,nombre,' . $planManto->id_plan_manto . ',id_plan_manto',
             'descripcion' => 'required|string|max:45|unique:plan_manto,descripcion,' . $planManto->id_plan_manto . ',id_plan_manto',
             'frecuencia_mes' => 'required|integer|min:1',
         ]);
+        $currentDateTime = Carbon::now('America/Guatemala');
+        $planManto->nombre = $request->input('nombre');
+        $planManto->descripcion = $request->input('descripcion');
+        $planManto->frecuencia_mes = $request->input('frecuencia_mes');
+        $planManto->updated_at = $currentDateTime;
+        $planManto->save();
 
-        $planManto->update($validatedData);
-
-        return redirect()->route('plan_manto.index')->with('success', 'Plan de mantenimiento actualizado exitosamente.');
+        return redirect()->route('planes.index')->with('success', 'Plan de mantenimiento actualizado exitosamente.');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(PlanManto $planManto)
+    public function destroy($hashedId)
     {
+        $id = $this->hashids->decode($hashedId)[0] ?? null;
+        if (!$id) {
+            abort(404);
+        }
+        $planManto = PlanManto::findOrFail($id);
         $planManto->delete();
 
-        return redirect()->route('plan_manto.index')->with('success', 'Plan de mantenimiento eliminado exitosamente.');
+        return redirect()->route('planes.index')->with('success', 'Plan de mantenimiento eliminado exitosamente.');
     }
 }
